+++ src/seat/keyboard/mod.rs
@@ -1,6 +1,3 @@
-#[rustfmt::skip]
-pub mod keysyms;
-
 use std::{
     convert::TryInto,
     env,
@@ -14,6 +11,10 @@ use std::{
     },
     time::Duration,
 };
+
+use xkeysym::KeyCode;
+#[doc(inline)]
+pub use xkeysym::Keysym;
 
 #[cfg(feature = "calloop")]
 use calloop::timer::{TimeoutAction, Timer};
@@ -137,7 +138,7 @@ pub trait KeyboardHandler: Sized {
         surface: &wl_surface::WlSurface,
         serial: u32,
         raw: &[u32],
-        keysyms: &[u32],
+        keysyms: &[Keysym],
     );
 
     /// The keyboard has left a surface.
@@ -248,8 +249,8 @@ pub struct KeyEvent {
 
     /// The interpreted symbol of the key.
     ///
-    /// This corresponds to one of the values in the [`keysyms`] module.
-    pub keysym: u32,
+    /// This corresponds to one of the assoiated values on the [`Keysym`] type.
+    pub keysym: Keysym,
 
     /// UTF-8 interpretation of the entered text.
     ///
@@ -584,7 +585,7 @@ where
                         .copied()
                         // We must add 8 to the keycode for any functions we pass the raw keycode into per
                         // wl_keyboard protocol.
-                        .map(|raw| guard.key_get_one_sym(raw + 8))
+                        .map(|raw| guard.key_get_one_sym(KeyCode::new(raw + 8)))
                         .collect::<Vec<_>>();
 
                     // Drop guard before calling user code.
@@ -618,7 +619,7 @@ where
                     if let Some(guard) = state_guard.as_ref() {
                         // We must add 8 to the keycode for any functions we pass the raw keycode into per
                         // wl_keyboard protocol.
-                        let keysym = guard.key_get_one_sym(key + 8);
+                        let keysym = guard.key_get_one_sym(KeyCode::new(key + 8));
                         let utf8 = if state == wl_keyboard::KeyState::Pressed {
                             let mut compose = udata.xkb_compose.lock().unwrap();
 
@@ -627,13 +628,15 @@ where
                                     xkb::FeedResult::Ignored => None,
                                     xkb::FeedResult::Accepted => match compose.status() {
                                         xkb::Status::Composed => compose.utf8(),
-                                        xkb::Status::Nothing => Some(guard.key_get_utf8(key + 8)),
+                                        xkb::Status::Nothing => {
+                                            Some(guard.key_get_utf8(KeyCode::new(key + 8)))
+                                        }
                                         _ => None,
                                     },
                                 },
 
                                 // No compose
-                                None => Some(guard.key_get_utf8(key + 8)),
+                                None => Some(guard.key_get_utf8(KeyCode::new(key + 8))),
                             }
                         } else {
                             None
@@ -675,7 +678,9 @@ where
                                         let key_repeats = state_guard
                                             .as_ref()
                                             .map(|guard| {
-                                                guard.get_keymap().key_repeats(event.raw_code + 8)
+                                                guard
+                                                    .get_keymap()
+                                                    .key_repeats(KeyCode::new(event.raw_code + 8))
                                             })
                                             .unwrap_or_default();
                                         if key_repeats {
@@ -792,15 +797,18 @@ where
                                     xkb::FeedResult::Ignored => None,
                                     xkb::FeedResult::Accepted => match compose.status() {
                                         xkb::Status::Composed => compose.utf8(),
-                                        xkb::Status::Nothing => {
-                                            Some(state.key_get_utf8(event.key.raw_code + 8))
-                                        }
+                                        xkb::Status::Nothing => Some(
+                                            state
+                                                .key_get_utf8(KeyCode::new(event.key.raw_code + 8)),
+                                        ),
                                         _ => None,
                                     },
                                 },
 
                                 // No compose.
-                                None => Some(state.key_get_utf8(event.key.raw_code + 8)),
+                                None => {
+                                    Some(state.key_get_utf8(KeyCode::new(event.key.raw_code + 8)))
+                                }
                             }
                         };